Move to managing hoyden, exgod under Morph.
This appears to be working out pretty well.

Alice is now a required user for all machines to manage remote
deployment requirements.

Probably should figure out a reasonable refactor in order to move
Alice into a "secrets" sort of thing, so that I'm not giving away user
info for free on git, but i'm not that worried about it.