Manage DO domains in Terraform!!
This is a nice step forward, for me. But there are some challenges
here I need to fix.

The first, and biggest one, is that DO rate limits API requests per
minute to ~250. If I'm doing even a small amount of changes and
iterating a lot I will run into issues with that just due to recrods;
each record pull appears to be its OWN CALL. God.

With that in mind, an easy next step seems like splitting the domain
management into their own submodules to prevent unnecessary calls.

For now, i'm just happy to have this managed in code :).