Configuration.nix pruning, tests.
Pantheon is actually really nice, but unusable for me on nixos because
of how I use ssh agent. KDE provides ksshaskpass as a total
replacement for pinentry, it seems, and Pantheon just uses pinentry.

On nixos, pinentry appears to be Fucked Up, and changing pinentry
binaries to other nixos packages doesn't result in anything besides
more of regular pinentry, which doesn't allow copy paste, which is a
NON STARTER.

FUcking linux.